% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles_df.R
\name{articles_df}
\alias{articles_df}
\title{Generates a tibble of news with title, description, publication date, url,  and image.}
\usage{
articles_df(x)
}
\arguments{
\item{x}{is the JSON file obtained from  NewsApi.org}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with title, description, publication date, url,  and image.
}
