% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_df.R
\name{data_df}
\alias{data_df}
\title{Generates a tibble of news with title, description, published date, url and image url}
\usage{
data_df(x)
}
\arguments{
\item{x}{is the parsed data}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with title, description, published date, url and image url
}
