% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything_df.R
\name{everything_df}
\alias{everything_df}
\title{Generates a tibble of news with author, title, description, url, imageUrl, publication date, and content}
\usage{
everything_df(x, verbose = FALSE)
}
\arguments{
\item{x}{is the JSON file obtained from  NewsApi.org}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with author, title, description, url, imageUrl, publication date, and content
}
