% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Generates a tibble of news with title, description, published date, active url and image}
\usage{
format_table(x)
}
\arguments{
\item{x}{is the parsed data}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with title, description, published date, active url and image
}
