% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_news.R
\name{get_data}
\alias{get_data}
\title{Creates a thinned Shapefile}
\usage{
get_data(APIkey = NULL)
}
\arguments{
\item{APIkey}{provided by the user to get the news headlines of USA}
}
\value{
parsed news
}
\description{
Creates a thinned Shapefile
}
