% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines_df.R
\name{headlines_df}
\alias{headlines_df}
\title{Generates a tibble of news with title, description, published date, url and image url}
\usage{
headlines_df(x)
}
\arguments{
\item{x}{is the APIkey to access news}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with title, description, published date, url and image url
}
