% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_news.R
\name{news_clean}
\alias{news_clean}
\title{Generates a tibble of news with title, description, published date and image link}
\usage{
news_clean(APIkey = NULL)
}
\arguments{
\item{APIkey}{is the APIkey to access news}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with title, description, published date and image link
}
