% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_news.R
\name{news_df}
\alias{news_df}
\title{Creates a thinned Shapefile}
\usage{
news_df(x, verbose = FALSE)
}
\arguments{
\item{parsed}{news}
}
\value{
a tibble of news
}
\description{
Creates a thinned Shapefile
}
