% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_to_df.R
\name{news_to_df}
\alias{news_to_df}
\title{Generates a tibble of news with title, description, published date, news url and image url}
\usage{
news_to_df(x)
}
\arguments{
\item{x}{is the APIkey to access news}
}
\value{
a tibble of news
}
\description{
Generates a tibble of news with title, description, published date, news url and image url
}
