% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{runshiny}
\alias{runshiny}
\title{Runs the shiny app that allows users to make search for articles and news headlines}
\usage{
runshiny()
}
\description{
Runs the shiny app that allows users to make search for articles and news headlines
}
