% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD:man/save_apikey.Rd
% Please edit documentation in R/Save_apikey.R
=======
% Please edit documentation in R/save_apikey.R
>>>>>>> 226b58cd5ba49a181b13211486bcda58821dcd77:man/Save_apikey.Rd
\name{save_apikey}
\alias{save_apikey}
\title{Saves the apikey to project .Renviron for future use}
\usage{
save_apikey(apikey = NULL)
}
\arguments{
\item{apikey}{is the APIkey obtained from Newsapi.org}
}
\description{
Saves the apikey to project .Renviron for future use
}
