% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_apikey.R
\name{save_apikey}
\alias{save_apikey}
\title{Saves the apikey to project .Renviron for future use}
\usage{
save_apikey(apikey = NULL)
}
\arguments{
\item{apikey}{is the APIkey obtained from Newsapi.org}
}
\description{
Saves the apikey to project .Renviron for future use
}
